convert DecksV16.count() to Int
V16 seems to use a long, but this isn't handled yet

+++ AnkiDroid/src/main/java/com/ichi2/libanki/DecksV16.kt
@@ -318,7 +318,7 @@ class DecksV16(private val col: Collection, private val decksBackend: DecksBacke
         this.save(deck)
     }
 
-    fun count(): Long {
+    fun count(): Int {
         return len(this.all_names_and_ids())
     }
